Convert ProjectForm component to .tsx
